requirements.txt
asyncpg>=0.29.0
PyJWT>=2.8.0

